Fix formatting, move file ext function to use case
